Move deploy init parsing test
