+++ tests/cli/test_deploy.py
@@ -90,6 +90,37 @@ def interactive_console(monkeypatch: pytest.MonkeyPatch):
     monkeypatch.setattr("readchar._posix_read.readchar", readchar)
 
 
+def test_deploy_init_field_value_can_contain_equals(
+    monkeypatch: pytest.MonkeyPatch, tmp_path: Path
+):
+    captured_inputs: dict[str, Any] = {}
+
+    def initialize_project(
+        name: str | None = None,
+        recipe: str | None = None,
+        inputs: dict[str, Any] | None = None,
+    ) -> list[str]:
+        captured_inputs.update(inputs or {})
+        return ["prefect.yaml"]
+
+    monkeypatch.setattr("prefect.deployments.initialize_project", initialize_project)
+
+    invoke_and_assert(
+        [
+            "deploy",
+            "init",
+            "--recipe",
+            "local",
+            "--field",
+            "api_key=aGVsbG8gd29ybGQ=",
+        ],
+        temp_dir=str(tmp_path),
+        expected_code=0,
+    )
+
+    assert captured_inputs == {"api_key": "aGVsbG8gd29ybGQ="}
+
+
 @pytest.fixture
 def project_dir(tmp_path: Path):
     with tmpchdir(str(tmp_path)):
